requirements.txt
youtube-transcript-api>=0.6.2
google-generativeai>=0.4.1
python-dotenv>=1.0.1
